Implement expanded view basic functionality. Separated ImageGallery into -Default and -Expanded React Components. Introduced new action and reducer to handle view selection
