+++ client/src/state/actions/defaultView.js
@@ -1,0 +1,5 @@
+export default () => {
+  return {
+    type: 'DEFAULT_VIEW',
+  };
+};
