+++ client/src/state/reducers/index.js
@@ -7,6 +7,7 @@ import size from './selectSize';
 import quantity from './selectQuantity';
 import cart from './addToCart';
 import related from './setRelated';
+import defaultView from './defaultView';
 
 const allReducers = {
   product, // object { productID: 16060, productInfo: { /products/:product_id }, styleInfo: { /products/:product_id/styles } }
@@ -16,6 +17,7 @@ const allReducers = {
   quantity, // number
   cart, // array of objects, each object representing a purchased product/style/size/qty
   related, // array of product_ids
+  defaultView, // boolean (true = default view)
 };
 
 export default combineReducers(allReducers);
